backend/requirements.txt
asgiref==3.4.1
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
Django==3.2.7
django-filter==21.1
django-mathfilters==1.0.0
django-widget-tweaks==1.4.9
idna==3.3
importlib-metadata==4.8.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pytz==2021.3
razorpay==1.2.0
regex==2021.11.10
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typed-ast==1.5.0
typing-extensions==4.0.0
urllib3==1.26.7
zipp==3.6.0
